requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.2
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-language==1.2.0
google-cloud-speech==1.1.0
googleapis-common-protos==1.6.0
grpcio==1.22.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.4
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
rsa==4.0
scipy==1.3.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4